backend/src/requirements.txt
pydantic
pydantic-settings
fastapi
python-multipart
pypdf
python-docx
langchain-experimental
langchain-aws
